+++ mapbox/libandroid-services/src/main/java/com/mapbox/services/android/navigation/v5/listeners/NewRouteProgressListener.java
@@ -1,4 +1,4 @@
-package com.mapbox.services.android.navigation.v5;
+package com.mapbox.services.android.navigation.v5.listeners;
 
 import com.mapbox.services.api.navigation.v5.RouteProgress;
